Phase 1: Split monolithic __init__.py into modular backend
- Extract Pydantic models to page_data.py (replaces dataclass Author + manual dicts)
- Extract DB helpers to internal_db.py (insert_comment, author_from_*, label helpers)
- Create router.py with datasette-plugin-router Router + check_permission decorator
- Move API handlers to routes/api.py using router decorators + Body() parsing
- Move page handlers to routes/pages.py
- Slim __init__.py from 850 to ~110 lines (hooks + content script injection only)
- Add datasette-plugin-router, datasette-vite, pydantic to dependencies
- Switch to dependency-groups for dev deps
- Update Justfile test recipe to support arguments
