+++ datasette_comments/routes/api.py
@@ -1,0 +1,548 @@
+from typing import Annotated, List
+from datasette import Response
+from datasette.utils import tilde_decode, tilde_encode
+from datasette.utils import await_me_maybe
+from datasette.plugins import pm
+from ulid import ULID
+import json
+
+from datasette_plugin_router import Body
+
+from ..router import router, check_permission
+from ..internal_db import (
+    insert_comment,
+    author_from_actor,
+    author_from_id,
+    author_from_request,
+    get_label_column,
+    get_label_for_row,
+)
+from ..page_data import (
+    ThreadNewRequest,
+    ThreadNewResponse,
+    CommentAddRequest,
+    OkResponse,
+    ThreadMarkResolvedRequest,
+    TableViewThreadsRequest,
+    TableViewThreadsResponse,
+    RowViewThreadsRequest,
+    RowViewThreadsResponse,
+    ReactionRequest,
+    AutocompleteMentionsResponse,
+    ActivitySearchResponse,
+)
+from .. import comment_parser
+
+
+@router.GET(
+    r"^/-/datasette-comments/api/thread/comments/(?P<thread_id>.*)$",
+    output=None,
+)
+@check_permission()
+async def thread_comments(thread_id: str, datasette=None, request=None):
+    results = await datasette.get_internal_database().execute(
+        """
+          select
+            id,
+            author_actor_id,
+            created_at,
+            (strftime('%s', 'now') - strftime('%s', created_at)) as created_duration_seconds,
+            contents,
+            (
+              select json_group_array(
+                json_object(
+                  'reactor_actor_id', reactor_actor_id,
+                  'reaction', reaction
+                )
+              )
+              from datasette_comments_reactions
+              where comment_id == datasette_comments_comments.id
+            ) as reactions
+          from datasette_comments_comments
+          where thread_id = ?
+          order by created_at
+        """,
+        (thread_id,),
+    )
+
+    actor_ids = set()
+    rows = []
+    for row in results:
+        actor_ids.add(row["author_actor_id"])
+        rows.append(dict(row))
+    actors = await datasette.actors_from_ids(list(actor_ids))
+    for row in rows:
+        author = author_from_actor(datasette, actors, row["author_actor_id"])
+        row["author"] = author.model_dump()
+
+        results = comment_parser.parse(row["contents"])
+        row["render_nodes"] = results.rendered
+        row["reactions"] = json.loads(row["reactions"]) if row["reactions"] else []
+    return Response.json({"ok": True, "data": rows})
+
+
+@router.POST(
+    r"^/-/datasette-comments/api/thread/new$",
+    output=ThreadNewResponse,
+)
+@check_permission(write=True)
+async def thread_new(
+    body: Annotated[ThreadNewRequest, Body()], datasette=None, request=None
+):
+    actor_id = request.actor.get("id")
+
+    type = body.type
+    database = body.database
+    table = body.table
+    column = body.column
+    rowids = body.rowids
+    comment = body.comment
+
+    # validate the target is good, depending on type
+    if type == "database":
+        if database is None:
+            return Response.json(
+                {"message": "target type database requires 'database' field"},
+                status=400,
+            )
+    elif type == "table":
+        if any(item is None for item in (database, table)):
+            return Response.json(
+                {
+                    "message": "target type table requires 'database' and 'table' fields"
+                },
+                status=400,
+            )
+    elif type == "row":
+        if any(item is None for item in (database, table, rowids)):
+            return Response.json(
+                {
+                    "message": "target type database requires 'database', 'table', and 'rowids' fields"
+                },
+                status=400,
+            )
+    elif type == "column":
+        if any(item is None for item in (database, table, column)):
+            return Response.json(
+                {
+                    "message": "target type column requires 'database', 'table', and 'column' fields"
+                },
+                status=400,
+            )
+        raise Exception("TODO column type not implemented")
+    elif type == "value":
+        if any(item is None for item in (database, table, column, rowids)):
+            raise Exception(
+                "target type value requires 'database', 'table', 'column', and 'rowids' fields"
+            )
+        raise Exception("TODO value type not implemented")
+    else:
+        raise Exception(f"target type '{type}' not supported")
+
+    # the urls input is a tilde-encoded string, so we split into individual primary keys here
+    rowids_decoded = None
+    if rowids is not None:
+        rowids_decoded = [tilde_decode(b) for b in rowids.split(",")]
+
+    id = str(ULID()).lower()
+
+    def db_thread_new(conn):
+        cursor = conn.cursor()
+        cursor.execute("begin")
+        params = {
+            "id": id,
+            "creator_actor_id": actor_id,
+            "target_type": type,
+            "target_database": database,
+            "target_table": table if type != "database" else None,
+            "target_column": column if type in ("column", "row", "value") else None,
+            "target_row_ids": (
+                json.dumps(rowids_decoded) if type in ("row", "value") else None
+            ),
+        }
+
+        cursor.execute(
+            """
+              insert into datasette_comments_threads(
+                id,
+                creator_actor_id,
+                target_type,
+                target_database,
+                target_table,
+                target_column,
+                target_row_ids
+              )
+              values (
+                :id,
+                :creator_actor_id,
+                :target_type,
+                :target_database,
+                :target_table,
+                :target_column,
+                :target_row_ids
+              );
+            """,
+            params,
+        )
+        thread_id = cursor.execute(
+            "select id from datasette_comments_threads where rowid = ?",
+            (cursor.lastrowid,),
+        ).fetchone()[0]
+
+        cursor.execute(*(insert_comment(thread_id, actor_id, comment)))
+        cursor.execute("commit")
+        return thread_id
+
+    try:
+        thread_id = await datasette.get_internal_database().execute_write_fn(
+            db_thread_new,
+            block=True,
+        )
+        return Response.json({"ok": True, "thread_id": thread_id})
+    except Exception as e:
+        raise e
+
+
+@router.POST(
+    r"^/-/datasette-comments/api/thread/comment/add$",
+    output=OkResponse,
+)
+@check_permission(write=True)
+async def comment_add(
+    body: Annotated[CommentAddRequest, Body()], datasette=None, request=None
+):
+    actor_id = request.actor.get("id")
+
+    await datasette.get_internal_database().execute_write(
+        *(insert_comment(body.thread_id, actor_id, body.contents)),
+        block=True,
+    )
+
+    return Response.json({"ok": True})
+
+
+@router.POST(
+    r"^/-/datasette-comments/api/threads/mark_resolved$",
+    output=OkResponse,
+)
+@check_permission(write=True)
+async def thread_mark_resolved(
+    body: Annotated[ThreadMarkResolvedRequest, Body()], datasette=None, request=None
+):
+    await datasette.get_internal_database().execute_write(
+        """
+            UPDATE datasette_comments_threads
+            SET resolved_at = datetime('now')
+            WHERE id = ?
+        """,
+        (body.thread_id,),
+        block=True,
+    )
+
+    return Response.json({"ok": True})
+
+
+@router.POST(
+    r"^/-/datasette-comments/api/threads/table_view$",
+    output=TableViewThreadsResponse,
+)
+@check_permission()
+async def table_view_threads(
+    body: Annotated[TableViewThreadsRequest, Body()], datasette=None, request=None
+):
+    database = body.database
+    table = body.table
+    rowids_encoded: List[str] = body.rowids
+    rowids: List[List[str]] = []
+    for rowid_encoded in rowids_encoded:
+        parts = [tilde_decode(b) for b in rowid_encoded.split(",")]
+        rowids.append(parts)
+
+    response = await datasette.get_internal_database().execute(
+        """
+          select id
+          from datasette_comments_threads
+          where target_type == 'table'
+            and target_database == ?1
+            and target_table == ?2
+            and not marked_resolved
+       """,
+        (database, table),
+    )
+    table_threads = [dict(row) for row in response.rows]
+
+    response = await datasette.get_internal_database().execute(
+        """
+          select
+            id,
+            target_row_ids
+          from datasette_comments_threads
+          where target_type == 'row'
+            and target_database == ?1
+            and target_table == ?2
+            and target_row_ids in (
+              select value
+              from json_each(?3)
+            )
+            and not marked_resolved
+       """,
+        (database, table, json.dumps(rowids)),
+    )
+    row_threads = [
+        {
+            "id": row["id"],
+            "rowids": "/".join(
+                map(lambda x: tilde_encode(x), json.loads(row["target_row_ids"]))
+            ),
+        }
+        for row in response.rows
+    ]
+
+    return Response.json(
+        {
+            "ok": True,
+            "data": {
+                "table_threads": table_threads,
+                "column_threads": [],
+                "row_threads": row_threads,
+                "value_threads": [],
+            },
+        }
+    )
+
+
+@router.POST(
+    r"^/-/datasette-comments/api/threads/row_view$",
+    output=RowViewThreadsResponse,
+)
+@check_permission()
+async def row_view_threads(
+    body: Annotated[RowViewThreadsRequest, Body()], datasette=None, request=None
+):
+    database = body.database
+    table = body.table
+    rowids_encoded: str = body.rowids
+    rowids = [tilde_decode(b) for b in rowids_encoded.split(",")]
+
+    response = await datasette.get_internal_database().execute(
+        """
+          select
+            id
+          from datasette_comments_threads
+          where target_type == 'row'
+            and target_database == ?1
+            and target_table == ?2
+            and target_row_ids = ?3
+            and not marked_resolved
+       """,
+        (database, table, json.dumps(rowids)),
+    )
+    row_threads = [row["id"] for row in response.rows]
+
+    return Response.json(
+        {
+            "ok": True,
+            "data": {
+                "row_threads": row_threads,
+            },
+        }
+    )
+
+
+@router.GET(
+    r"^/-/datasette-comments/api/reactions/(?P<comment_id>.*)$",
+    output=None,
+)
+@check_permission()
+async def reactions(comment_id: str, datasette=None, request=None):
+    results = await datasette.get_internal_database().execute(
+        """
+          SELECT
+            reactor_actor_id,
+            reaction
+          FROM datasette_comments_reactions
+          WHERE comment_id == :comment_id
+        """,
+        {"comment_id": comment_id},
+    )
+    return Response.json([dict(row) for row in results.rows])
+
+
+@router.POST(
+    r"^/-/datasette-comments/api/reaction/add$",
+    output=OkResponse,
+)
+@check_permission(write=True)
+async def reaction_add(
+    body: Annotated[ReactionRequest, Body()], datasette=None, request=None
+):
+    id = str(ULID()).lower()
+    reactor_actor_id = request.actor.get("id")
+
+    await datasette.get_internal_database().execute_write(
+        """
+          INSERT INTO datasette_comments_reactions(
+            id,
+            comment_id,
+            reactor_actor_id,
+            reaction
+          )
+          VALUES (
+            :id,
+            :comment_id,
+            :reactor_actor_id,
+            :reaction
+          )
+        """,
+        {
+            "id": id,
+            "comment_id": body.comment_id,
+            "reactor_actor_id": reactor_actor_id,
+            "reaction": body.reaction,
+        },
+        block=True,
+    )
+    return Response.json({"ok": True})
+
+
+@router.POST(
+    r"^/-/datasette-comments/api/reaction/remove$",
+    output=OkResponse,
+)
+@check_permission(write=True)
+async def reaction_remove(
+    body: Annotated[ReactionRequest, Body()], datasette=None, request=None
+):
+    reactor_actor_id = request.actor.get("id")
+
+    await datasette.get_internal_database().execute_write(
+        """
+          DELETE FROM datasette_comments_reactions
+          WHERE comment_id = :comment_id
+            AND reactor_actor_id = :reactor_actor_id
+            AND reaction = :reaction
+        """,
+        {
+            "comment_id": body.comment_id,
+            "reactor_actor_id": reactor_actor_id,
+            "reaction": body.reaction,
+        },
+        block=True,
+    )
+    return Response.json({"ok": True})
+
+
+@router.GET(
+    r"^/-/datasette-comments/api/autocomplete/mentions$",
+    output=AutocompleteMentionsResponse,
+)
+@check_permission(write=True)
+async def autocomplete_mentions(datasette=None, request=None):
+    prefix = request.args.get("prefix")
+    suggestions = []
+    for users in pm.hook.datasette_comments_users(datasette=datasette):
+        for user in await await_me_maybe(users):
+            username = user.get("username")
+            if username and username.startswith(prefix):
+                author = await author_from_id(datasette, user.get("id"))
+                suggestions.append(
+                    {
+                        "username": user.get("username"),
+                        "author": author.model_dump(),
+                    }
+                )
+    return Response.json({"suggestions": suggestions})
+
+
+@router.GET(
+    r"^/-/datasette-comments/api/activity_search$",
+    output=ActivitySearchResponse,
+)
+@check_permission()
+async def activity_search(datasette=None, request=None):
+    search_comments = request.args.get("searchComments")
+    author = request.args.get("author")
+    database = request.args.get("database")
+    table = request.args.get("table")
+    is_resolved = request.args.get("isResolved") == "1"
+    contains_tag = request.args.getlist("containsTag")
+
+    WHERE = "1"
+    params = []
+
+    if search_comments:
+        WHERE += " AND comments.contents LIKE printf('%%%s%%', ?)"
+        params.append(search_comments)
+
+    if author:
+        for users in pm.hook.datasette_comments_users(datasette=datasette):
+            for user in await await_me_maybe(users):
+                if user.get("username") == author:
+                    WHERE += " AND comments.author_actor_id = ?"
+                    params.append(user.get("id"))
+                    break
+
+    if database:
+        WHERE += " AND threads.target_database = ?"
+        params.append(database)
+
+    if table:
+        WHERE += " AND threads.target_table = ?"
+        params.append(table)
+
+    WHERE += f" AND {'' if is_resolved else 'NOT'} threads.marked_resolved"
+
+    for tag in contains_tag:
+        if not tag:
+            continue
+        WHERE += " AND ? in (select value from json_each(comments.hashtags))"
+        params.append(tag)
+
+    sql = f"""
+          SELECT
+            comments.author_actor_id,
+            comments.contents,
+            comments.created_at,
+            (strftime('%s', 'now') - strftime('%s', comments.created_at)) as created_duration_seconds,
+            threads.target_type,
+            threads.target_database,
+            threads.target_table,
+            threads.target_row_ids,
+            threads.target_column
+          FROM datasette_comments_comments AS comments
+          LEFT JOIN datasette_comments_threads AS threads ON threads.id = comments.thread_id
+          WHERE {WHERE}
+          ORDER BY comments.created_at DESC
+          LIMIT 100;
+    """
+    results = await datasette.get_internal_database().execute(sql, params)
+    data = [dict(row) for row in results.rows]
+
+    actor_ids = set(row["author_actor_id"] for row in data)
+    actors = await datasette.actors_from_ids(actor_ids)
+
+    for row in data:
+        row_author = author_from_actor(datasette, actors, row["author_actor_id"])
+        row["author"] = row_author.model_dump()
+
+        label_column = await get_label_column(
+            datasette, row["target_database"], row["target_table"]
+        )
+        if label_column:
+            try:
+                rowids = json.loads(row["target_row_ids"])
+            except Exception:
+                row["target_label"] = None
+                continue
+
+            target_label = await get_label_for_row(
+                datasette.databases[row["target_database"]],
+                row["target_table"],
+                label_column,
+                rowids,
+            )
+            row["target_label"] = target_label
+        else:
+            row["target_label"] = None
+
+    return Response.json({"data": data})
