+++ datasette_comments/routes/pages.py
@@ -1,0 +1,14 @@
+from datasette import Response
+
+from ..router import router, check_permission
+
+
+@router.GET(r"^/-/datasette-comments/activity$")
+@check_permission()
+async def activity_view(datasette=None, request=None):
+    return Response.html(
+        await datasette.render_template(
+            "activity_view.html",
+            request=request,
+        )
+    )
