+++ pyproject.toml
@@ -4,7 +4,7 @@ version = "0.1.1.a3"
 description = "A Datasette plugin for commenting on tables, rows, and values"
 readme = "README.md"
 authors = [{name = "Alex Garcia"}]
-dependencies = ["datasette>=1.0a21", "python-ulid", "sqlite-migrate==0.1b0"]
+dependencies = ["datasette>=1.0a21", "python-ulid", "sqlite-migrate==0.1b0", "datasette-plugin-router>=0.0.1a2", "datasette-vite>=0.0.1a2", "pydantic>=2"]
 license = "Apache-2.0"
 classifiers=[
     "Framework :: Datasette"
@@ -20,8 +20,8 @@ CI = "https://github.com/datasette/datasette-comments/actions"
 [project.entry-points.datasette]
 comments = "datasette_comments"
 
-[project.optional-dependencies]
-test = ["pytest", "pytest-asyncio", "pytest-playwright", "syrupy"]
+[dependency-groups]
+dev = ["pytest", "pytest-asyncio", "pytest-playwright", "syrupy"]
 
 [tool.setuptools.packages.find]
 namespaces = false
